requirements.txt
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
sqlalchemy>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
folium>=0.15.0
shapely>=1.8.5
geopy>=2.4.0
pyyaml>=6.0
loguru>=0.7.0
schedule>=1.2.0
anthropic>=0.20.0
jinja2>=3.1.0
aiohttp>=3.9.0
tqdm>=4.64.0
python-dateutil>=2.8.0
numpy>=1.24.0
tenacity>=8.2.0
branca>=0.6.0
